casl integrate, only existence

+++ src/models/inventory/existence/existence.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards, Req, HttpException, ForbiddenException, ParseIntPipe, Query } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards, Req, HttpException, ForbiddenException, ParseIntPipe, Query, HttpCode, HttpStatus } from '@nestjs/common';
 import { ExistenceService } from './existence.service';
 import { CreateExistenceDto } from './dto/create-existence.dto';
 import { UpdateExistenceDto } from './dto/update-existence.dto';
@@ -6,9 +6,8 @@ import { ApiParam, ApiQuery, ApiTags } from '@nestjs/swagger';
 import { JwtAuthGuard } from '../../authentication/guards/jwt-auth.guard';
 import { IRequestWithUser } from 'src/models/interfaces/models.interface';
 import { ForbiddenError, subject } from '@casl/ability';
-import { create } from 'lodash';
 import { CaslAbilityFactory } from '../../authentication/authorization/casl/casl-ability.factory';
-import { Existence } from './entities/existence.entity';
+import { resources } from '../../../helpers/consts';
 
 @Controller('existence')
 @ApiTags('Existencias')
@@ -19,29 +18,38 @@ export class ExistenceController {
   ) { }
 
   @Post()
+  // @UseGuards(PoliciesGuard)
+  // @CheckPolicies(new CreateExistenceCheck())
   @UseGuards(JwtAuthGuard)
   create(@Body() createSubsidiaryExistenceDto: CreateExistenceDto, @Req() request: IRequestWithUser) {
     try {
       const { user } = request;
-
       const ability = this.caslAbilityFactory.createForUser(user);
-      ForbiddenError.from(ability).throwUnlessCan('create', subject('Existence', createSubsidiaryExistenceDto));
-      return this.existenceService.create(createSubsidiaryExistenceDto);
+      ForbiddenError.from(ability).setMessage(resources.access_denied).throwUnlessCan('create', subject('Existence', createSubsidiaryExistenceDto));
+      // return this.existenceService.create(createSubsidiaryExistenceDto);
     } catch (error) {
       if (error instanceof ForbiddenError) {
         throw new ForbiddenException(`${error.message}`);
       }
+      throw new HttpException(`${error.message}`, error.status);
     }
   }
 
-  @UseGuards(JwtAuthGuard)
   @Get()
+  // @UseGuards(PoliciesGuard)
+  // @CheckPolicies(new ReadExistenceCheck())
+  @UseGuards(JwtAuthGuard)
   findAll(@Req() request: IRequestWithUser) {
     try {
       const { user } = request;
       const ability = this.caslAbilityFactory.createForUser(user);
-      ForbiddenError.from(ability).setMessage('Sin acceso a este recurso').throwUnlessCan('read', subject('Existence', user));
-      return this.existenceService.findAll(user.subsidiaryId);
+      ForbiddenError.from(ability).setMessage(resources.access_denied).throwUnlessCan('read', 'Existence');
+
+      let isSadmin: boolean;
+      if (['sadmin'].includes(user.role.name)) isSadmin = true;
+
+      return this.existenceService.findAll({ subsidiaryId: user.subsidiaryId, isSadmin: isSadmin });
+
     } catch (error) {
       if (error instanceof ForbiddenError) {
         throw new ForbiddenException(`${error.message}`);
@@ -50,29 +58,54 @@ export class ExistenceController {
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.existenceService.findOne(+id);
+  // @UseGuards(PoliciesGuard)
+  // @CheckPolicies(new ReadExistenceCheck())
+  @UseGuards(JwtAuthGuard)
+  async findOne(@Param('id', ParseIntPipe) id: number, @Req() request: IRequestWithUser) {
+    try {
+      const { user } = request;
+
+      const existence = await this.existenceService.findOne(id);
+
+      if (!existence) throw new HttpException(`recurso ${resources.not_found}`, 404);
+
+      const ability = this.caslAbilityFactory.createForUser(user);
+      ForbiddenError.from(ability).setMessage(resources.access_denied).throwUnlessCan('read', subject('Existence', existence));
+
+      return existence;
+
+    } catch (error) {
+      if (error instanceof ForbiddenError) {
+        throw new ForbiddenException(`${error.message}`);
+      }
+      throw new HttpException(`${error.message}`, error.status);
+    }
   }
 
-  // @Get('')
-  // findOneBy(@Param('id') id: string) {
-  //   return this.subsidiaryExistenceService.findOne(+id);
-  // }
-
   @Patch(':id')
+  // @UseGuards(PoliciesGuard)
+  // @CheckPolicies(new PatchExistenceCheck())
+  @UseGuards(JwtAuthGuard)
   update(@Param('id') id: string, @Body() updateSubsidiaryExistenceDto: UpdateExistenceDto) {
     return this.existenceService.update(+id, updateSubsidiaryExistenceDto);
   }
 
   @ApiParam({ name: 'id', type: Number })
-  @ApiQuery({ name: 'soft', type: Boolean, required: false })
+  @ApiQuery({ name: 'soft', type: Boolean})
   @Delete('remove/:id')
+  // @UseGuards(PoliciesGuard)
+  // @CheckPolicies(new DeleteExistenceCheck())
+  @HttpCode(HttpStatus.NO_CONTENT)
   @UseGuards(JwtAuthGuard)
-  remove(@Param('id', ParseIntPipe) id: number, @Query('soft') soft: boolean, @Req() request: IRequestWithUser) {
+  async remove(@Param('id', ParseIntPipe) id: number, @Query('soft') soft: boolean, @Req() request: IRequestWithUser) {
     try {
       const { user } = request;
+      const existence = await this.existenceService.findOne(id);
+      if (!existence) throw new HttpException(`recurso ${resources.not_found}`, 404);
+      
       const ability = this.caslAbilityFactory.createForUser(user);
-      ForbiddenError.from(ability).setMessage('No puede eliminar este recurso, no posee permisos').throwUnlessCan('delete', 'Existence');
+      ForbiddenError.from(ability).throwUnlessCan('delete', subject('Existence', existence));
+      
       return this.existenceService.remove(id, soft);
 
     } catch (error) {
@@ -80,7 +113,31 @@ export class ExistenceController {
         throw new ForbiddenException(`${error.message}`);
       }
       throw new HttpException(`${error.message}`, error.status);
+    }
+  }
 
+  @Patch('restore/:id')
+  // @UseGuards(PoliciesGuard)
+  // @CheckPolicies(new PatchExistenceCheck())
+  @UseGuards(JwtAuthGuard)
+  async restore(id: number, @Req() request: IRequestWithUser) {
+    try {
+      const { user } = request;
+
+      const ability = this.caslAbilityFactory.createForUser(user);
+      const existence = await this.existenceService.findOne(id);
+
+      ForbiddenError.from(ability).setMessage(resources.access_denied).throwUnlessCan('read', subject('Existence', existence));
+
+      if (!existence) throw new HttpException(`recurso ${resources.not_found}`, 404);
+
+      this.existenceService.restore(id);
+
+    } catch (error) {
+      if (error instanceof ForbiddenError) {
+        throw new ForbiddenException(`${error.message}`);
+      }
+      throw new HttpException(`${error.message}`, error.status);
     }
   }
 }
